requirements.txt
tensorflow
PyYAML
numpy
tqdm
